requirements.txt
requests
pendulum
quantiphy
python-dotenv
